requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
redis==5.0.1
sqlalchemy[asyncio]==2.0.23

# Telegram Bot
python-telegram-bot==20.7
aiohttp==3.9.1
pyngrok==7.0.0

# RPA/Browser automation
selenium==4.16.0
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4

# External APIs
openai==1.3.7
requests==2.31.0
httpx==0.25.2
PyJWT==2.8.0

# Address validation
dadata==25.7.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
colorlog==6.8.0
celery==5.3.4
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0

# Data processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# PDF Generation
reportlab==4.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring (for production)
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
